Remove unnecessary step for the image
The current version of the image is explicitly installing rvm,
setting up user, then installing ruby with bundler and then install
everything using bundler which should not be necessary for our use
case.

This commit change this, now it usages a ruby slim image as base, so
we don't have to setup anything for ruby explicitly. This also install
gems directly and keeps the version number as variable.

Related: #19